Move face reversal code to dedicate module
